Changes to fix some serious failures: * Updated main.c to conform to the new kernel behavior wrt   requires/provides.  Main.c submits PEOS_EVENT_REQUIRES when it sees   an action is blocked, and PEOS_EVENT_PROVIDES when an action is   pending. * Changed the way proc_graph.c finds the 'dot' executable.  The path   to 'dot' is now set at compile time, in the Makefile. * Modified makedot to orient the graph horizontally, rather than   vertically; it's also forced into a 10'' wide page, which makes it   easier to see the whole graph, but harder to read.

+++ ui/web2/cgi-bin/test_do.sh
@@ -1,7 +1,8 @@
 #!/bin/sh
 echo -n "$0..."
 
-rm proc_table.dat
+test -e proc_table.dat && rm proc_table.dat
+
 create_testtable
 
 export QUERY_STRING="0+test_script=run"
